move node-creation related files to node dir
