+++ client/lib/strata/stratum/view/node/nodeColorCSS.js
