+++ client/lib/strata/stratum/view/node/nodeTemplate.js
